requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
SQLAlchemy==2.0.*
alembic==1.14.*
psycopg2-binary==2.9.*
gunicorn==23.*
requests==2.32.*
